reorganize deneg py pkg
