+++ deneg/deneg/examples.py
