+++ deneg/deneg/utils.py
